requirements.txt
absl-py==2.1.0
asgiref==3.8.1
attrs==25.1.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
cloudinary==1.42.1
dj-database-url==2.3.0
dj-rest-auth==7.0.1
Django==5.1.5
django-allauth==65.3.1
django-cloudinary-storage==0.3.0
django-cors-headers==4.6.0
django-countries==7.6.1
django-filter==24.3
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
dm-tree==0.1.9
docstring_parser==0.16
etils==1.11.0
fsspec==2025.2.0
googleapis-common-protos==1.66.0
gunicorn==23.0.0
idna==3.10
immutabledict==4.2.1
importlib_resources==6.5.2
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml-dtypes==0.3.2
namex==0.0.8
optree==0.14.0
packaging==24.2
pandas==2.2.3
pillow==11.1.0
promise==2.3
psutil==6.1.1
psycopg2==2.9.10
pyarrow==19.0.0
Pygments==2.19.1
PyJWT==2.10.1
python-dateutil==2.9.0.post0
pytz==2025.1
pyuca==1.2
requests==2.32.3
rest-framework-simplejwt==0.0.2
rich==13.9.4
setuptools==75.8.0
simple-parsing==0.1.7
six==1.17.0
sqlparse==0.5.3
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow-datasets==4.9.7
tensorflow-metadata==1.16.1
tensorflow-recommenders==0.7.2
toml==0.10.2
tqdm==4.67.1
tzdata==2025.1
urllib3==2.3.0
Werkzeug==3.1.3
wheel==0.45.1
whitenoise==6.8.2
zipp==3.21.0
